+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed 
+
+- Bump QUDT version to [2.1.37](https://github.com/qudt/qudt-public-repo/releases/tag/v2.1.37)
+
 ## [6.5.0] - 2024-03-17
 
 ### Added