+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.7.0] - 2024-04-04
+
 ### Changed
 
 - Bump QUDT version to [2.1.37](https://github.com/qudt/qudt-public-repo/releases/tag/v2.1.37)
@@ -277,7 +279,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/qudtlib/qudtlib-java/compare/v6.5.0...HEAD
+[Unreleased]: https://github.com/qudtlib/qudtlib-java/compare/v6.7.0...HEAD
+[6.7.0]: https://github.com/qudtlib/qudtlib-java/compare/v6.5.0...v6.7.0
 [6.5.0]: https://github.com/qudtlib/qudtlib-java/compare/v6.4.3...v6.5.0
 [6.4.3]: https://github.com/qudtlib/qudtlib-java/compare/v6.4.2...v6.4.3
 [6.4.2]: https://github.com/qudtlib/qudtlib-java/compare/v6.4.1...v6.4.2